% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{buildDateQuery}
\alias{buildDateQuery}
\title{build date query}
\usage{
buildDateQuery(
  dateValue,
  dateFieldName = "gameDate",
  date_side = c("equality", "less_than", "less_than_equals", "greater_than",
    "greater_than_equals")
)
}
\arguments{
\item{dateValue}{Date query value as a date object}

\item{dateFieldName}{character date field name.}

\item{date_side}{character with possible values \code{"less_than"},
\code{"less_than_equals"}. \code{"equality"}, \code{"greater_than_equals"}
and \code{"greater_than"}.}
}
\value{
character query in JSON type
}
\description{
build date query by searching date from midnight to midnight.
}
\examples{
m = mongo("Fixture", "Examples")
## find all fixtures in Feb 3, 2017
m$count(buildDateQuery(as.Date("2017-02-03"))) ## 16
}
