% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{buildFieldStatsQuery}
\alias{buildFieldStatsQuery}
\title{Build a query to compute field statistics}
\usage{
buildFieldStatsQuery(field, sub_field)
}
\arguments{
\item{field}{character, name of the field to compute statistics for.}

\item{sub_field}{character (optional), sub field. If provided the
\code{field} name will be treated as main field and nested list
structure is assumed.}
}
\value{
character, a JSON query object.
}
\description{
Build a query to compute field statistics, currently max, min,
average, and standard deviation.
}
