% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{buildIntervalQuery}
\alias{buildIntervalQuery}
\title{build date interval query}
\usage{
buildIntervalQuery(
  date_min,
  date_max,
  dateFieldName = "gameDate",
  start_type = c("large", "strict"),
  end_type = c("large", "strict")
)
}
\arguments{
\item{date_min, date_max}{Date objects left and right hand side of the interval}

\item{dateFieldName}{character date field name.}

\item{start_type, end_type}{character with possible values \code{large} and
\code{strict}. Should the equality be large or strict.}
}
\value{
character query in JSON type.
}
\description{
build date interval query
}
\examples{
m = mongo("Fixture", "Examples")
## find all fixtures in Feb 3, 2017 (large) and Feb 8, 2017 (strict)
qr <- buildIntervalQuery(as.Date("2019-02-03"), as.Date("2019-02-08"),
                         start_type = "large", end_type = "strict")
m$count(qr) ## 37
}
