% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{buildMixedIntervalQuery}
\alias{buildMixedIntervalQuery}
\title{build a mixed query with date intervals and normal fields}
\usage{
buildMixedIntervalQuery(
  keys,
  values,
  type = c("equality", "not_equals", "less_than", "less_than_equals", "greater_than",
    "greater_than_equals", "exists", "intervals"),
  date_min,
  date_max,
  dateFieldName = "gameDate",
  start_type = c("large", "strict"),
  end_type = c("large", "strict")
)
}
\arguments{
\item{keys}{character vector keys}

\item{values}{list same length as keys with the fields values.}

\item{type}{character query operator. Either of length 1 (and in this case it
will be recycled to be \code{length(values)}) or of same length as keys
if different type should be applied to different keys. Accepted values
are \code{"equality"} (for all types of values), \code{"not_equals"},
\code{"less_than"}, \code{"less_than_equals"}, \code{"greater_than"},
\code{"greater_than_equals"} (for numerical values only), \code{"exists"}
(for logiical values) and \code{"intervals"} (for numerical values only;
large bounds will be applied)}

\item{date_min, date_max}{Date objects left and right hand side of
the interval}

\item{dateFieldName}{character date field name.}

\item{start_type, end_type}{character with possible vslues
\code{large} and \code{strict}. Should the equality be large or
strict.}
}
\value{
character query in JSON type
}
\description{
build a mixed query with date intervals and normal fields
}
