% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{buildObjectQuery}
\alias{buildObjectQuery}
\title{build query for objectId field}
\usage{
buildObjectQuery(objectId)
}
\arguments{
\item{objectId}{character objectId value}
}
\value{
character query test strings
}
\description{
build query for objectId field
}
\details{
This query will request the \code{_id} field.
}
