% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{buildQuery}
\alias{buildQuery}
\title{build a query for mongoDB request}
\usage{
buildQuery(keys, values, type = "equality")
}
\arguments{
\item{keys}{character vector keys}

\item{values}{list same length as keys with the fields values.}

\item{type}{character query operator. Either of length 1 (and in this case it
will be recycled to be \code{length(values)}) or of same length as keys
if different type should be applied to different keys. Accepted values
are \code{"equality"} (for all types of values), \code{"not_equals"},
\code{"less_than"}, \code{"less_than_equals"}, \code{"greater_than"},
\code{"greater_than_equals"} (for numerical values only), \code{"exists"}
(for logiical values) and \code{"intervals"} (for numerical values only;
large bounds will be applied)}
}
\value{
character JSON type query.
}
\description{
build a query for mongoDB request
}
\examples{
m = mongo("Fixture", "Examples")
## find all fixtures from competitionId 8 in seasonId 2017
keys = list(competitionId = 8, seasonId = 2017)
m$count(buildQuery(names(keys), keys)) ## 380
}
