% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{buildRandomQuery}
\alias{buildRandomQuery}
\title{build query to extract random documents from the database}
\usage{
buildRandomQuery(qr, nSample = 1)
}
\arguments{
\item{qr}{character a valid query object.}

\item{nSample}{numeric nombre of documents to extract}
}
\value{
a character query to pass to the aggregate method.
}
\description{
build query to extract random documents from the database
}
\details{
This query is to be used with the \code{aggregate()} method.
}
\examples{
m = mongo("Fixture", "Examples")
## find all fixtures in competition 8 seasonId 2017, played after Feb 3, 2017
keys = list(competitionId = 8, seasonId = 2017)
qr = buildQuery(names(keys), keys)
m$aggregate(buildRandomQuery(qr, 2))$gameId
}
