% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveObject.R
\name{insert_object}
\alias{insert_object}
\alias{insert_object_inDB}
\alias{read_object_fromDB}
\title{Insert an RDS file in mongoDB}
\usage{
insert_object_inDB(
  objectPath,
  remotename,
  metadata,
  db = "models",
  prefix = "fs",
  verbose = TRUE
)

read_object_fromDB(remotename, db = "models", prefix = "fs", verbose = TRUE)
}
\arguments{
\item{objectPath}{character, path to the RDS file to insert.}

\item{remotename}{character, the name the file will be known as
within the GridFS.}

\item{metadata}{character (in JSON format), extra info to store.}

\item{db}{character, database name.}

\item{prefix}{identifies the GridFS \code{.chunks} and
\code{.files} data collections.}

\item{verbose}{logical, should progress messages be printed to the
screen ?}
}
\value{
logical giving the insertion status (2019-07-17 @Tarak: I
    believe that the value is not logical, see Details above and
    delete the wrong one); it also has a data.frame attribute
    \code{att} with the function arguments \code{objectPath},
    \code{remotename}, \code{db} and \code{prefix}.
}
\description{
Insert an RDS file in mongoDB using the GridFS technique.

Read a GridFS inserted object from mongoDB.
}
\details{
Connects to a GridFS prefix and adds the RDS file to it. The name of
the file must have extention \code{".RDS"}, an error is thrown
otherwise.

The return value is as from the write() method for GridFS objects. In
addition, it has attribute \code{"att"}, which is a data frame with
one row containing some arguments from the call (\code{remotename},
\code{db}, \code{prefix}) and the GridFS object returned by
\code{mongolite::gridfs}.
}
