% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{mongo_atlas}
\alias{mongo_atlas}
\title{convenient function to connect to mongo-atlas cluster}
\usage{
mongo_atlas(
  db,
  collection,
  username,
  password,
  cluster_urls = atlas_urls_,
  replicaset = atlas_cluster_name_,
  use_ssl = TRUE,
  authSource = "admin"
)
}
\arguments{
\item{db}{character, name of the mongo database.}

\item{collection}{character, name of the mongo collection in \code{db}.}

\item{username}{character, credentials to connect to mongo atlas.}

\item{password}{character, credentials to connect to mongo atlas.}

\item{cluster_urls}{character, vector clusters urls together with connection
port.}

\item{replicaset}{character, main cluster url}

\item{use_ssl}{logical ssl connection ?}

\item{authSource}{character (optional), do we have an authentification source
key?}
}
\description{
convenient function to connect to mongo-atlas cluster
}
\details{
The user can go to the mongo-Atlas website to find the different urls
required.  He can be inspired by the shell command example.
}
\examples{
\dontrun{
m2 <- mongo_atlas("ExternalData", "Player")
print(m2$count())
}

}
