% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{mongo_connect}
\alias{mongo_connect}
\title{wrapper function to connect to mongo}
\usage{
mongo_connect(
  db_name,
  collection,
  user = "",
  password = "",
  url = "localhost",
  replicaset = NULL,
  port = 27017,
  use_ssl = FALSE,
  authSource = NULL,
  connection_delay = NULL,
  mongo_base = "mongodb://",
  retry_writes = NULL,
  majority = NULL
)
}
\arguments{
\item{db_name}{character, name of the mongo database.}

\item{collection}{character, name of the mongo collection in
\code{db_name}.}

\item{user, password}{character, credentials to connect to mongo
atlas.}

\item{url}{character, vector clusters urls together with connection
port.}

\item{replicaset}{character, main cluster url.}

\item{port}{numeric, connection port.}

\item{use_ssl}{logical, ssl connection ?}

\item{authSource}{character (optional), do we have an
authentification source key?}

\item{connection_delay}{numeric should we apply a socket-timout ? should be
provided in ms.}
}
\value{
the result of running the command \code{mongolite::mongo()}
}
\description{
wrapper function to connect to mongo
}
\details{
The function assists the user to connect to a mongo-server using
either local host or a replica set.
}
\examples{
lconfig_local <- list(url = "localhost", replicaset = NA_character_,
                      port =  27017, db_name = "ExternalData",
                      user = "", password =  "")
lconfig_local$collection <- "Player"
m <- do.call(mongo_connect, lconfig_local)
print(m$count())

}
