% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate.R
\name{mongo_replicate}
\alias{mongo_replicate}
\title{mongo replicate cluster driver}
\usage{
mongo_replicate(
  db,
  collection,
  username,
  password,
  cluster_urls = aws_urls_,
  replicaset = aws_cluster_name_,
  port = 27017L,
  use_ssl = FALSE,
  authSource = NULL,
  connection_delay = NULL,
  mongo_base = "mongodb://",
  retry_writes = NULL,
  majority = NULL
)
}
\arguments{
\item{db}{character, name of the mongo database.}

\item{collection}{character, name of the mongo collection in \code{db}.}

\item{username, password}{character, credentials to connect to mongo atlas.}

\item{cluster_urls}{character, vector clusters urls together with connection
port.}

\item{replicaset}{character, main cluster url}

\item{port}{numeric, connection port.}

\item{use_ssl}{logical ssl connection ?}

\item{authSource}{character (optional), do we have an authentification source
key?}

\item{connection_delay}{numeric should we apply a socket-timout ? should be
provided in ms.}
}
\value{
the result of running the command \code{mongolite::mongo()}
}
\description{
mongo replicate cluster driver
}
\details{
Assist the user to connect to the replicate mongo cluster.
}
\examples{
\dontrun{
m1 <- mongo_replicate("InternalDataBookie", "allFeatures")
print(m1$count())
}

}
