% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MongoTools.R
\docType{package}
\name{mongoTools}
\alias{mongoTools}
\title{Functions to Interact with MongoDB.}
\description{
A set of tools to assist the user requesting data from mongoDB. We cover in
particular query generating and cluster connection. The package is meant to
complete mongolite.
}
\details{
In order to run the examples, make sure you restore your mongoDB with the
dump shipped with this package. For example, this will create database
'Examples':

    \code{mongorestore -d Examples mongoTools/inst/extdata/dump/Examples/}
}
\examples{
admin <- mongo(db = "admin")
## one of the databases shown by this should be 'Examples' (see above)
admin$run('{"listDatabases":1}')
}
