% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{prod_mongo_con}
\alias{prod_mongo_con}
\title{test mongo connection in a production like environment}
\usage{
prod_mongo_con(collection, lconfig)
}
\arguments{
\item{collection}{character collection name}

\item{lconfig}{list database configuration passed as a list}
}
\value{
numeric the connection total count
}
\description{
test mongo connection in a production like environment
}
