% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{test_mongo_con}
\alias{test_mongo_con}
\title{Test mongo connection in a production like environment}
\usage{
test_mongo_con(collection, lconfig)
}
\arguments{
\item{collection}{character, name of the collection.}

\item{lconfig}{list, further arguments for
\code{\link{mongo_connect}}.}
}
\value{
numeric, number of records in the collection
}
\description{
Test mongo connection in a production like environment. Connects to
the collection and returns the number of records in it.
}
\examples{
test_mongo_con("Fixture", lconfig = list(db_name = "Examples")) # 76292
}
