% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/query.R
\name{form_update_query}
\alias{form_update_query}
\alias{form_update_query.logical}
\alias{form_update_query.numeric}
\alias{form_update_query.character}
\alias{form_update_query.Date}
\alias{form_update_query.POSIXct}
\alias{update_query}
\alias{buildUpdateQuery}
\title{generic function to form the appropriate update query}
\usage{
form_update_query(object, key, ...)

\method{form_update_query}{logical}(object, key, ...)

\method{form_update_query}{numeric}(object, key, ...)

\method{form_update_query}{character}(object, key, ...)

\method{form_update_query}{Date}(object, key, ...)

\method{form_update_query}{POSIXct}(object, key, ...)

buildUpdateQuery(keys, values)
}
\arguments{
\item{object}{the object we create query for (the value).}

\item{key}{character field name.}

\item{...}{further arguments for methods.}

\item{keys}{character vector keys}

\item{values}{list same length as keys with the fields values.}
}
\value{
character, the query
}
\description{
form a query for the appropriate update query

build update query block for logical type

build update query block for numeric type

build update query block for character type

build update query block for Date type

build update query block for \code{POSIXct} type

build a query useful for update operations.
}
